+++ CacheMoney/frontend/src/components/SplashView.js
