+++ CacheMoney/frontend/src/components/Transaction.js
@@ -1,7 +1,7 @@
 import React from "react";
 
-function Transaction() {
-	return <></>;
+function Transaction(props) {
+	return <>Date - Description - Type - Amount - new balance?</>;
 }
 
 export default Transaction;
